% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{gv.to.svg}
\alias{gv.to.svg}
\title{Create a svg file from a gv (Graphiviz) file or text}
\usage{
gv.to.svg(gv, use.memoise = TRUE, to.clipboard = is.null(out.file),
  out.file = NULL)
}
\arguments{
\item{gv}{filename or source code of graphiviz diagram}

\item{use.memoise}{if TRUE memoise results}

\item{to.clipboard}{copy SVG source code to cliboard}

\item{out.file}{if not NULL save to outfile}
}
\description{
Create a svg file from a gv (Graphiviz) file or text
}

